# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "招聘中心系统 API 文档",
        "title": "招聘中心 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/resumes": {
            "post": {
                "description": "创建用户简历信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "简历管理"
                ],
                "summary": "创建简历",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "简历信息",
                        "name": "resume",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateResumeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Resume"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/resumes/access-status": {
            "put": {
                "description": "更新用户简历的访问状态（公开/隐藏）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "简历管理"
                ],
                "summary": "更新简历访问状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "访问状态",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateResumeStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/resumes/my": {
            "get": {
                "description": "获取当前登录用户的简历信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "简历管理"
                ],
                "summary": "获取用户简历",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Resume"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/resumes/share/{token}": {
            "get": {
                "description": "根据分享token获取简历信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "简历管理"
                ],
                "summary": "根据分享token获取简历",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分享Token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Resume"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/resumes/upload": {
            "post": {
                "description": "上传用户简历文件",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "简历管理"
                ],
                "summary": "上传简历文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "简历文件",
                        "name": "resume",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/resumes/working-status": {
            "put": {
                "description": "更新用户简历的工作状态（在职/离职）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "简历管理"
                ],
                "summary": "更新简历工作状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "工作状态",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateResumeStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/resumes/{id}": {
            "put": {
                "description": "更新用户简历信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "简历管理"
                ],
                "summary": "更新简历",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "简历ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新简历信息",
                        "name": "resume",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateResumeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/resumes/{id}/favorite": {
            "post": {
                "description": "切换简历的收藏状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "简历管理"
                ],
                "summary": "收藏/取消收藏简历",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "简历ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/resumes/{id}/stats": {
            "get": {
                "description": "获取简历的查看和收藏统计信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "简历管理"
                ],
                "summary": "获取简历统计信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "简历ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "integer",
                                                "format": "int64"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/resumes/{id}/view": {
            "post": {
                "description": "记录简历查看行为",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "简历管理"
                ],
                "summary": "查看简历收藏相关信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "简历ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.Resume": {
            "type": "object",
            "properties": {
                "accessStatus": {
                    "description": "1: 隐藏, 2: 公开",
                    "type": "integer"
                },
                "attachments": {
                    "description": "新增附件关联",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ResumeAttachment"
                    }
                },
                "avatar": {
                    "type": "string"
                },
                "birthday": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "educations": {
                    "description": "关联",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Education"
                    }
                },
                "email": {
                    "type": "string"
                },
                "expectedCity": {
                    "type": "string"
                },
                "expectedJob": {
                    "type": "string"
                },
                "experience": {
                    "type": "integer"
                },
                "gender": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "introduction": {
                    "type": "string"
                },
                "jobStatus": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Project"
                    }
                },
                "shareToken": {
                    "description": "用于分享简历的唯一标识",
                    "type": "string"
                },
                "skills": {
                    "type": "string"
                },
                "status": {
                    "description": "1: 正常, 0: 删除",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                },
                "workExperiences": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.WorkExperience"
                    }
                },
                "workingStatus": {
                    "description": "1: 在职, 2: 离职",
                    "type": "integer"
                }
            }
        },
...
        "request.CreateResumeRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "phone"
            ],
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "birthday": {
                    "type": "string"
                },
                "educations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.EducationRequest"
                    }
                },
                "email": {
                    "type": "string"
                },
                "expectedCity": {
                    "type": "string"
                },
                "expectedJob": {
                    "type": "string"
                },
                "experience": {
                    "type": "integer"
                },
                "gender": {
                    "type": "integer"
                },
                "introduction": {
                    "type": "string"
                },
                "jobStatus": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.ProjectRequest"
                    }
                },
                "skills": {
                    "type": "string"
                },
                "workExperiences": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.WorkExperienceRequest"
                    }
                }
            }
        },
        "request.EducationRequest": {
            "type": "object",
            "required": [
                "school"
            ],
            "properties": {
                "degree": {
                    "type": "string"
                },
                "endTime": {
                    "type": "string"
                },
                "major": {
                    "type": "string"
                },
                "school": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
...
        "request.UpdateResumeRequest": {
            "type": "object",
            "required": [
                "data",
                "module"
            ],
            "properties": {
                "data": {
                    "description": "更新数据"
                },
                "module": {
                    "description": "更新模块",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.UpdateResumeModuleType"
                        }
                    ]
                }
            }
        },
        "request.UpdateResumeStatusRequest": {
            "type": "object",
            "required": [
                "resumeId",
                "status",
                "userId"
            ],
            "properties": {
                "resumeId": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/errors.ErrorCode"
                        }
                    ]
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应信息",
                    "type": "string"
                }
            }
        }
    }
}